Move UI and IO impls into cli package
