--- pkg/clui/interface.go
@@ -1,10 +1,0 @@
-package clui
-
-type UI interface {
-	Section(msg string)
-	Subsection(msg string)
-	ItemSuccess(msg string)
-	ItemSkipped(msg string)
-	ItemFailure(msg string)
-	Confirm(msg string) (bool, error)
-}
